+++ includes/key.h
@@ -6,7 +6,7 @@
 /*   By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/14 14:53:28 by kipark            #+#    #+#             */
-/*   Updated: 2022/11/18 16:57:26 by kipark           ###   ########seoul.kr  */
+/*   Updated: 2022/11/18 18:41:55 by kipark           ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,8 +28,6 @@
 # define RIGHT_ARROR_KEY		123
 # define LEFT_ARROR_KEY			124
 # define WALL_DIST	 			3
-
-
 
 int		key_press(int keycode, t_game *param);
 void	rot_view(t_player *player, double rot);
