+++ includes/struct.h
@@ -6,7 +6,7 @@
 /*   By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/14 17:23:00 by kipark            #+#    #+#             */
-/*   Updated: 2022/11/18 15:41:40 by kipark           ###   ########seoul.kr  */
+/*   Updated: 2022/11/18 17:09:15 by kipark           ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
