+++ srcs/game_view/game_draw.c
@@ -6,7 +6,7 @@
 /*   By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/16 17:06:09 by kipark            #+#    #+#             */
-/*   Updated: 2022/11/18 14:42:59 by kipark           ###   ########seoul.kr  */
+/*   Updated: 2022/11/18 18:01:13 by kipark           ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
